Support date-time separator in datetime()
As I discussed in #54, I, and the rest of my team, often find myself writing
`$dt->ymd . ' ' . $dt->hms` because $dt->datetime uses the ISO8601 "T"
separator, which is somewhat unpleasant and not average-user-friendly.

I feel this simple change would be really helpful.

I implemented it in `datetime()` only, resulting in a tiny bit of duplicated
code, because it doesn't seem to make sense to be able to pass a separator to
`iso8601()`, when it's ISO8601 which specifies "T" as a separator, and thus
you'd be telling `iso8601()` to generate a datetime string that's not in
ISO8601 format.