Add ignore_missing_spans parameter to DateTime->new()
Without this parameter, if the given time does not exist in the given time
zone due to DST, DateTime->new() will throw an exception.

With this parameter, DateTime->new() will instead think the offset for
the time is the pre-DST offset. This will cause the time to jump forward
one hour, while still ending up with the proper DST offset and is_dst flag
at the end.

For example:

  my $dt = DateTime->new(
    year   => 2007,
    month  => 3,
    day    => 11,
    hour   => 2,
    minute => 30,
    second => 0,
    time_zone => 'America/New_York',
    ignore_missing_spans => 1,
  );

  print $dt->offset;  # -14400 (-4 (DST))
  print "$dt";        # "2018-03-11T03:30:00"
  print $dt->ist_dst; # 1

This follows https://tools.ietf.org/html/rfc5545#page-34:

  If the local time described does not occur (when
  changing from standard to daylight time), the DATE-TIME value is
  interpreted using the UTC offset before the gap in local times.
  Thus, TZID=America/New_York:20070311T023000 indicates March 11,
  2007 at 3:30 A.M. EDT (UTC-04:00), one hour after 1:30 A.M. EST
  (UTC-05:00).